screensaverEnable: Rework code to eliminate race condition(s)
The screensaverEnable flag was ignored most of the time because the thread
running the proximity code got stopped so fast almost every time sending
the stop() signal even before the thread was setting the state to 'running'
in the start() method.
Insert a new flag earlyStopReq which handles exactly this case.
This led to a new race condition concerning the stop/destruction mechanism
of the thread: QThread: Destroyed while thread is still running
This is solved by a modification of the connections between InputNotifier
and IdleHelper and a rearrangement of the cleanUp method/finished() signal.

Signed-off-by: Peter Fink <pfink@christ-es.de>