main.qml: add onRenderProcessTerminated handling
If the renderProcess dies or is killed, then we reload the
WebEngineView. This restarts the renderProcess.

There are two renderProcesses visible in the system. If the
child dies then this works fine. If the parent process dies
then this won't help at all, because then there is no way
to detect it and get back to a sane state.

Signed-off-by: Peter Fink <pfink@christ-es.de>